+++ src/theme/edgeConfig.ts
@@ -11,6 +11,8 @@ export const edgeConfig: AppConfig = {
   configName: 'edge',
   darkTheme: edgeDark,
   defaultWallets: ['BTC', 'ETH', 'LTC', 'BCH', 'DASH'],
+  forceCloseUrlIos: 'https://support.apple.com/en-us/HT201330 ',
+  forceCloseUrlAndroid: 'https://support.google.com/android/answer/9079646?hl=en',
   knowledgeBase: 'https://help.edge.app/support/home',
   lightTheme: edgeLight,
   notificationServers: ['https://notif1.edge.app'],